requirements_cpu.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-limiter==0.1.6
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.0
h11==0.16.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.0
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMatting==1.1.14
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
redis==6.1.0
referencing==0.36.2
rembg==2.0.66
requests==2.32.3
rpds-py==0.25.1
s3transfer==0.13.0
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tifffile==2025.5.21
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
